requirements.txt
autopep8==1.4.4
certifi==2019.9.11
chardet==3.0.4
Click==7.0
dominate==2.4.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.2
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.17.2
ortools==7.3.7083
pep8==1.7.1
protobuf==3.10.0
psycopg2==2.8.3
pycodestyle==2.5.0
python-dotenv==0.10.3
requests==2.22.0
six==1.12.0
SQLAlchemy==1.3.9
urllib3==1.25.6
visitor==0.1.3
Werkzeug==0.16.0
WTForms==2.2.1
